+++ resolve/lusol/lusol_precision.f90
@@ -41,5 +41,5 @@ module lusol_precision
   implicit none
   public
 
-  integer(4),   parameter :: ip = int64, rp = real64
+  integer(4),   parameter :: ip = int32, rp = real64
 end module lusol_precision
